Fix R-tree bugs: node_choose containment, id type, allocator, stack buffer
- node_choose() checked the root bounding box (&rtree->box) instead of the
  i-th child's bbox (RTREE_NODE_BBOX_N(node, i)), causing all insertions to
  go into child 0 and degenerating the tree into a linked list.

- Normalize id type to int everywhere. The node struct stored int64 ids but
  internal functions (node_insert, add_answer, node_swap) and rtree_search
  used int, silently truncating. An in-memory index won't exceed INT_MAX
  entries, so int is sufficient and saves 256 bytes per node.

- node_free/rtree_free used free() but nodes/tree are allocated with
  palloc0(). Use pfree() consistently to match the allocator.

- node_swap() and unioned_area() used STBox as a stack buffer for memcpy,
  relying on the implicit invariant that STBox is the largest bbox type.
  Use char buf[sizeof(STBox)] to make this explicit and self-documenting.

+++ meos/include/meos.h
@@ -274,7 +274,7 @@ extern RTree *rtree_create_tstzspan();
 extern RTree *rtree_create_tbox();
 extern RTree *rtree_create_stbox();
 extern void rtree_free(RTree *rtree);
-extern void rtree_insert(RTree *rtree, void *box, int64 id);
+extern void rtree_insert(RTree *rtree, void *box, int id);
 extern int *rtree_search(const RTree *rtree,const void *query, int *count);
 
 /*****************************************************************************
